DOCSP-51825 Standardize Struct Tag Usage ex
